Some basic components of the hash table implementation in C.
In hash_table.h, the key-value pairs are stored in a struct. As of now,
I still don't quite understand the reason why the author use two
pointers to the key and the value. Hopefully I can wrap my head around
it soon.

The hash table data structure stores the `size` of that hash table, a
`count` variable that stores an int that indicates how full it is. The
hash table also stores an array of pointers to items (Which is in fact
a pointer to pointers of items).

Next, we move to the functions in `hash_table.c`. I create a function to
create new `ht_item`, the function is static because it cannot be called
from anywhere other than the hash table itself. This function returns a
pointer to the `ht_item`.  This function allocates a chunk of memory the size of an ht_item,
and saves a copy of the strings k and v in the new chunk of memory. This
is my first time using the malloc function in C and I think I can
understand it now :)

The function `ht_new` initialises a new hash table.`size` defines how
many `ht_item` we can store, it is 53 for now, but I will be able to
resize soon. The array of items is initialised with the C function `calloc`
, which fill the allocated memory with Null bytes ( which means that the
array is empty)